+++ tests/PureField/FieldFactoryTest.php
@@ -342,8 +342,7 @@ final class FieldFactoryTest extends TestCase
 
         $expected = <<<HTML
             <div>
-            <fieldset>
-            </fieldset>
+            <fieldset></fieldset>
             </div>
             HTML;
 
@@ -362,8 +361,7 @@ final class FieldFactoryTest extends TestCase
 
         $expected = <<<HTML
             <span>
-            <fieldset>
-            </fieldset>
+            <fieldset></fieldset>
             </span>
             HTML;